*: improve cluster conditions
