+++ pkg/apis/etcd/v1beta2/cluster.go
@@ -16,9 +16,7 @@ package v1beta2
 
 import (
 	"errors"
-	"fmt"
 	"strings"
-	"time"
 
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -222,180 +220,4 @@ func (c *ClusterSpec) Cleanup() {
 	}
 
 	c.Version = strings.TrimLeft(c.Version, "v")
-}
-
-type ClusterPhase string
-
-const (
-	ClusterPhaseNone     ClusterPhase = ""
-	ClusterPhaseCreating              = "Creating"
-	ClusterPhaseRunning               = "Running"
-	ClusterPhaseFailed                = "Failed"
-)
-
-type ClusterCondition struct {
-	Type ClusterConditionType `json:"type"`
-
-	Reason string `json:"reason"`
-
-	TransitionTime string `json:"transitionTime"`
-}
-
-type ClusterConditionType string
-
-const (
-	ClusterConditionReady = "Ready"
-
-	ClusterConditionRemovingDeadMember = "RemovingDeadMember"
-
-	ClusterConditionRecovering = "Recovering"
-
-	ClusterConditionScalingUp   = "ScalingUp"
-	ClusterConditionScalingDown = "ScalingDown"
-
-	ClusterConditionUpgrading = "Upgrading"
-)
-
-type ClusterStatus struct {
-	// Phase is the cluster running phase
-	Phase  ClusterPhase `json:"phase"`
-	Reason string       `json:"reason"`
-
-	// ControlPuased indicates the operator pauses the control of the cluster.
-	ControlPaused bool `json:"controlPaused"`
-
-	// Condition keeps ten most recent cluster conditions
-	Conditions []ClusterCondition `json:"conditions"`
-
-	// Size is the current size of the cluster
-	Size int `json:"size"`
-	// Members are the etcd members in the cluster
-	Members MembersStatus `json:"members"`
-	// CurrentVersion is the current cluster version
-	CurrentVersion string `json:"currentVersion"`
-	// TargetVersion is the version the cluster upgrading to.
-	// If the cluster is not upgrading, TargetVersion is empty.
-	TargetVersion string `json:"targetVersion"`
-
-	// BackupServiceStatus is the status of the backup service.
-	// BackupServiceStatus only exists when backup is enabled in the
-	// cluster spec.
-	BackupServiceStatus *BackupServiceStatus `json:"backupServiceStatus,omitempty"`
-}
-
-type MembersStatus struct {
-	// Ready are the etcd members that are ready to serve requests
-	// The member names are the same as the etcd pod names
-	Ready []string `json:"ready,omitempty"`
-	// Unready are the etcd members not ready to serve requests
-	Unready []string `json:"unready,omitempty"`
-}
-
-func (cs *ClusterStatus) IsFailed() bool {
-	if cs == nil {
-		return false
-	}
-	return cs.Phase == ClusterPhaseFailed
-}
-
-func (cs *ClusterStatus) SetPhase(p ClusterPhase) {
-	cs.Phase = p
-}
-
-func (cs *ClusterStatus) PauseControl() {
-	cs.ControlPaused = true
-}
-
-func (cs *ClusterStatus) Control() {
-	cs.ControlPaused = false
-}
-
-func (cs *ClusterStatus) UpgradeVersionTo(v string) {
-	cs.TargetVersion = v
-}
-
-func (cs *ClusterStatus) SetVersion(v string) {
-	cs.TargetVersion = ""
-	cs.CurrentVersion = v
-}
-
-func (cs *ClusterStatus) SetReason(r string) {
-	cs.Reason = r
-}
-
-func (cs *ClusterStatus) AppendScalingUpCondition(from, to int) {
-	c := ClusterCondition{
-		Type:           ClusterConditionScalingUp,
-		Reason:         scalingReason(from, to),
-		TransitionTime: time.Now().Format(time.RFC3339),
-	}
-	cs.appendCondition(c)
-}
-
-func (cs *ClusterStatus) AppendScalingDownCondition(from, to int) {
-	c := ClusterCondition{
-		Type:           ClusterConditionScalingDown,
-		Reason:         scalingReason(from, to),
-		TransitionTime: time.Now().Format(time.RFC3339),
-	}
-	cs.appendCondition(c)
-}
-
-func (cs *ClusterStatus) AppendRecoveringCondition() {
-	c := ClusterCondition{
-		Type:           ClusterConditionRecovering,
-		TransitionTime: time.Now().Format(time.RFC3339),
-	}
-	cs.appendCondition(c)
-}
-
-func (cs *ClusterStatus) AppendUpgradingCondition(to string, member string) {
-	reason := fmt.Sprintf("upgrading cluster member %s version to %v", member, to)
-
-	c := ClusterCondition{
-		Type:           ClusterConditionUpgrading,
-		Reason:         reason,
-		TransitionTime: time.Now().Format(time.RFC3339),
-	}
-	cs.appendCondition(c)
-}
-
-func (cs *ClusterStatus) AppendRemovingDeadMember(name string) {
-	reason := fmt.Sprintf("removing dead member %s", name)
-
-	c := ClusterCondition{
-		Type:           ClusterConditionRemovingDeadMember,
-		Reason:         reason,
-		TransitionTime: time.Now().Format(time.RFC3339),
-	}
-	cs.appendCondition(c)
-}
-
-func (cs *ClusterStatus) SetReadyCondition() {
-	c := ClusterCondition{
-		Type:           ClusterConditionReady,
-		TransitionTime: time.Now().Format(time.RFC3339),
-	}
-
-	if len(cs.Conditions) == 0 {
-		cs.appendCondition(c)
-		return
-	}
-
-	lastc := cs.Conditions[len(cs.Conditions)-1]
-	if lastc.Type == ClusterConditionReady {
-		return
-	}
-	cs.appendCondition(c)
-}
-
-func (cs *ClusterStatus) appendCondition(c ClusterCondition) {
-	cs.Conditions = append(cs.Conditions, c)
-	if len(cs.Conditions) > 10 {
-		cs.Conditions = cs.Conditions[1:]
-	}
-}
-
-func scalingReason(from, to int) string {
-	return fmt.Sprintf("Current cluster size: %d, desired cluster size: %d", from, to)
 }
